+++ Rogue.Types/Rectangle.cs
@@ -2,12 +2,12 @@
 {
     public class Rectangle
     {
-        public int X { get; set; }
+        public float X { get; set; }
 
-        public int Y { get; set; }
+        public float Y { get; set; }
 
-        public int Width { get; set; }
+        public float Width { get; set; }
 
-        public int Height { get; set; }
+        public float Height { get; set; }
     }
 }
